Added echelon form solver
Doesn't always work in specific situations such as when there's no solutions, and it's left in a weird decimal form. Apart from that it's okay I think!
If you wanna solve in more dimensions that what the ASCII converter thing would allow just comment out that bit of code, it only writes the variables into a classic linear equation format.